requirements.txt
kafka
kafka-python-ng
pyspark
redis
fastapi
uvicorn
cassandra-driver
